# hotel-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  hotel-backend_internal_models.CreateRoomRequest:
    properties:
      floor:
        type: integer
      room_number:
        type: string
      room_type_id:
        type: integer
      status:
        $ref: '#/definitions/hotel-backend_internal_models.RoomStatus'
    required:
    - floor
    - room_number
    - room_type_id
    type: object
  hotel-backend_internal_models.CreateRoomTypeRequest:
    properties:
      amenities:
        type: string
      base_price_per_night:
        type: number
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    required:
    - base_price_per_night
    - capacity
    - name
    type: object
...
  hotel-backend_internal_models.RoomStatus:
    enum:
    - available
    - occupied
    - cleaning
    - maintenance
    type: string
    x-enum-varnames:
    - RoomStatusAvailable
    - RoomStatusOccupied
    - RoomStatusCleaning
    - RoomStatusMaintenance
...
  hotel-backend_internal_models.UpdateRoomRequest:
    properties:
      floor:
        type: integer
      is_active:
        type: boolean
      room_number:
        type: string
      room_type_id:
        type: integer
      status:
        $ref: '#/definitions/hotel-backend_internal_models.RoomStatus'
    type: object
  hotel-backend_internal_models.UpdateRoomStatusRequest:
    properties:
      status:
        $ref: '#/definitions/hotel-backend_internal_models.RoomStatus'
    required:
    - status
    type: object
  hotel-backend_internal_models.UpdateRoomTypeRequest:
    properties:
      amenities:
        type: string
      base_price_per_night:
        type: number
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@hotel-system.com
    name: API Support
  description: This is the API documentation for the Hotel Management System.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Hotel Management System API
  version: "1.0"
paths:
...
  /room-types:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all room categories
      produces:
      - application/json
      responses:
        "200":
          description: Room types retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all room types
      tags:
      - Room Types
    post:
      consumes:
      - application/json
      description: Add a new room category (Admin only)
      parameters:
      - description: Room type details
        in: body
        name: room_type
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.CreateRoomTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Room type created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create room type
      tags:
      - Room Types
  /room-types/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a room type by ID (Admin only)
      parameters:
      - description: Room Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Room type deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid room type ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete room type
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete room type
      tags:
      - Room Types
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific room type
      parameters:
      - description: Room Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Room type retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid room type ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Room type not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get room type by ID
      tags:
      - Room Types
    put:
      consumes:
      - application/json
      description: Modify an existing room type's details (Admin only)
      parameters:
      - description: Room Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Room type update data
        in: body
        name: room_type
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.UpdateRoomTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Room type updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Room type not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update room type
      tags:
      - Room Types
  /rooms:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all rooms with optional filters for status,
        floor, and room_type_id
      parameters:
      - description: Room Status (available, occupied, cleaning, maintenance)
        in: query
        name: status
        type: string
      - description: Floor Number
        in: query
        name: floor
        type: integer
      - description: Room Type ID
        in: query
        name: room_type_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rooms retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all rooms
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Add a new room to the hotel inventory (Admin only)
      parameters:
      - description: Room data
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Room created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new room
      tags:
      - Rooms
  /rooms/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a room from hotel inventory (Admin only)
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Room deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid room ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete room
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete room
      tags:
      - Rooms
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific room by its ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Room retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid room ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Room not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get room by ID
      tags:
      - Rooms
    put:
      consumes:
      - application/json
      description: Modify room details such as floor, type, status, or keycard (Admin
        only)
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Room update data
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Room updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Room not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update room
      tags:
      - Rooms
  /rooms/{id}/status:
    patch:
      consumes:
      - application/json
      description: Change operational status of a room (available, occupied, cleaning,
        maintenance)
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update payload
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.UpdateRoomStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Room status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update room status
      tags:
      - Rooms
  /rooms/available:
    get:
      consumes:
      - application/json
      description: Search for available rooms matching given date range, optional
        room type, and guest capacity
      parameters:
      - description: Check-in Date (YYYY-MM-DD)
        in: query
        name: check_in
        required: true
        type: string
      - description: Check-out Date (YYYY-MM-DD)
        in: query
        name: check_out
        required: true
        type: string
      - description: Room Type ID
        in: query
        name: room_type_id
        type: integer
      - description: Minimum capacity
        in: query
        name: capacity
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Available rooms retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid dates or query parameters
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Check room availability
      tags:
      - Rooms
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"